+++ examples/shape/customNode/demo/lineChart.js
@@ -143,7 +143,7 @@ const data = {
       x: 150,
       y: 150,
       label: 'Line1',
-      shape: 'circleLine',
+      type: 'circleLine',
       anchorPoints: [
         [ 0, 0.5 ], [ 1, 0.5 ]
       ],
@@ -160,7 +160,7 @@ const data = {
       x: 500,
       y: 150,
       label: 'Line2',
-      shape: 'circleLine',
+      type: 'circleLine',
       anchorPoints: [
         [ 0, 0.5 ], [ 1, 0.5 ]
       ],