+++ examples/shape/label/demo/labelLen1.js
@@ -44,7 +44,7 @@ const graph = new G6.Graph({
   width,
   height,
   defaultNode: {
-    shape: 'rect',
+    type: 'rect',
     style: {
       fill: '#DEE9FF',
       stroke: '#5B8FF9'