+++ examples/tree/dendrogram/demo/basicDendrogram.js
@@ -29,7 +29,7 @@ fetch('https://gw.alipayobjects.com/os/antvdemo/assets/data/algorithm-category.j
         }
       },
       defaultEdge: {
-        shape: 'cubic-horizontal',
+        type: 'cubic-horizontal',
         style: {
           stroke: '#A3B1BF'
         }